+++ sdk/azurestackhci/azure-mgmt-azurestackhci/generated_tests/test_azure_stack_hci_updates_operations_async.py
@@ -137,3 +137,17 @@ class TestAzureStackHCIUpdatesOperationsAsync(AzureMgmtRecordedTestCase):
 
         # please add some check logic here by yourself
         # ...
+
+    @RandomNameResourceGroupPreparer(location=AZURE_LOCATION)
+    @recorded_by_proxy_async
+    async def test_updates_begin_prepare(self, resource_group):
+        response = await (
+            await self.client.updates.begin_prepare(
+                resource_group_name=resource_group.name,
+                cluster_name="str",
+                update_name="str",
+            )
+        ).result()  # call '.result()' to poll until service return final result
+
+        # please add some check logic here by yourself
+        # ...