+++ sdk/containerservice/azure-mgmt-containerservice/generated_samples/snapshots_delete.py
@@ -36,6 +36,6 @@ def main():
     )
 
 
-# x-ms-original-file: 2026-01-02-preview/SnapshotsDelete.json
+# x-ms-original-file: 2026-02-01/SnapshotsDelete.json
 if __name__ == "__main__":
     main()